Started working on adjacent cost reduction
There's a possible rabbit hole here.  Should the logic that says "yes, you can use this" also account for if there's an open, adjacent place on the board that will fit the polyomino?

Checking all that feels like it might be a pretty intense scope creep.  We'll stick with the system being a little dumber than that and see how it feels.